antigravity/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pydantic==2.10.0
pillow==11.0.0
transformers==4.47.0
torch==2.5.1
exifread==3.0.0
python-dotenv==1.0.1
